Refactor project to be consumable via GitHub API
