+++ forward-message-multiple/.env
@@ -1,0 +1,2 @@
+# A list of numbers [in E.164 format](https://support.twilio.com/hc/en-us/articles/223183008-Formatting-International-Phone-Numbers) you want to forward incoming messages to, separated by commas
+FORWARDING_NUMBERS=+12223334444,+491761234567
