+++ forward-message-multiple/package.json
@@ -1,0 +1,3 @@
+{
+  "dependencies": {}
+}
