+++ forward-message-sendgrid/.env
@@ -1,0 +1,6 @@
+# Your SendGrid API key
+SENDGRID_API_KEY=
+# The email address to forward the message to
+TO_EMAIL_ADDRESS=
+# The email address that SendGrid should send the email from
+FROM_EMAIL_ADDRESS=
