+++ forward-message-sparkpost/.env
@@ -1,0 +1,6 @@
+# Your SparkPost API key
+SPARKPOST_API_KEY=
+# The email address to forward the message to
+TO_EMAIL_ADDRESS=
+# The email address that SparkPost should send the email from
+FROM_EMAIL_ADDRESS=
