+++ forward-message-sparkpost/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "got": "^6.7.1"
+  }
+}
