+++ hunt/.env
@@ -1,0 +1,4 @@
+# REQUIRED: A comma separated list of numbers [in E.164 format](https://support.twilio.com/hc/en-us/articles/223183008-Formatting-International-Phone-Numbers) that you want to dial in order
+PHONE_NUMBERS=+12223334444,+491761234567
+# A URL to redirect the call to if none of the numbers answer. If this is not supplied then the call will just hang up once it has exhausted all the options
+FINAL_URL=
